refactor: ajout d'un useState pour controler l'ouverture et la fermeture de la modal
